alert utility mixin
Mixin created for alert utility classes. some styling is condition in this mixin. if we have icon in the start of alert we should have to set "true" to $hasIcons as a parameter when we @include this mixin  and if we need alert close icon we have to set "true" to  $closeBtn true as a parameter when @include this mixin. These true parameter will generate 2 different classes for each icon.